+++ src/main/java/com/github/alexthe666/iceandfire/entity/ai/DeathWormAIFindSandTarget.java
@@ -51,20 +51,20 @@ public class DeathWormAIFindSandTarget extends EntityAIBase {
     public BlockPos findSandTarget() {
         if (this.mob.getAttackTarget() == null || this.mob.getAttackTarget().isDead) {
             List<BlockPos> sand = new ArrayList<>();
-            if(this.mob.isTamed() && this.mob.getWormHome() != null){
+            if (this.mob.isTamed() && this.mob.getWormHome() != null) {
                 range = 25;
-                for (int x = (int) this.mob.getWormHome().getX() - range; x < (int) this.mob.getWormHome().getX() + range; x++) {
-                    for (int y = (int)this.mob.getWormHome().getY() - range; y < (int) this.mob.getWormHome().getY() + range; y++) {
-                        for (int z = (int) this.mob.getWormHome().getZ() - range; z < (int) this.mob.getWormHome().getZ() + range; z++) {
+                for (int x = this.mob.getWormHome().getX() - range; x < this.mob.getWormHome().getX() + range; x++) {
+                    for (int y = this.mob.getWormHome().getY() - range; y < this.mob.getWormHome().getY() + range; y++) {
+                        for (int z = this.mob.getWormHome().getZ() - range; z < this.mob.getWormHome().getZ() + range; z++) {
                             if (this.mob.world.getBlockState(new BlockPos(x, y, z)).getMaterial() == Material.SAND && isDirectPathBetweenPoints(this.mob, this.mob.getPositionVector(), new Vec3d(x, y, z))) {
                                 sand.add(new BlockPos(x, y, z));
                             }
                         }
                     }
                 }
-            }else{
+            } else {
                 for (int x = (int) this.mob.posX - range; x < (int) this.mob.posX + range; x++) {
-                    for (int y = (int)this.mob.posY - range; y < (int) this.mob.posY + range; y++) {
+                    for (int y = (int) this.mob.posY - range; y < (int) this.mob.posY + range; y++) {
                         for (int z = (int) this.mob.posZ - range; z < (int) this.mob.posZ + range; z++) {
                             if (this.mob.world.getBlockState(new BlockPos(x, y, z)).getMaterial() == Material.SAND && isDirectPathBetweenPoints(this.mob, this.mob.getPositionVector(), new Vec3d(x, y, z))) {
                                 sand.add(new BlockPos(x, y, z));
